hamidreza_esmailbeyk/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.2
exceptiongroup==1.3.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==25.0
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
pycodestyle==2.12.1
PyJWT==2.9.0
pytest==8.3.5
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
zipp==3.20.2
coverage==7.2.7
